review: fix code-review findings on #6037/#6040/#6041 batch
- Fix mismatched/asymmetric backticks in RebuildIndexStatement's
  post-drop exhaustion message.
- Document why DatabaseChecker's new .withMaxAttempts(LOCK_MAX_ATTEMPTS)
  can hold the outer file lock noticeably longer per attempt than before.
- Note the worst-case-latency trade-off (minutes rather than ~28s under
  sustained contention before giving up) in docs/release-26.9.1.md, since
  it's an operator-visible behavior change under load.
- Extract the corruptRecordTypeByte test helper and the lock-holding
  thread (previously two near-identical private classes) into shared,
  reusable TestHelper members, and switch CheckDatabaseFixPreservesIndexMetadataTest,
  DatabaseCheckerIndexRebuildLockContentionTest and RebuildIndexLockExhaustionTest
  to use them instead of their own copies.
